+++ ErsatzTV.Infrastructure/Data/Repositories/EmbyMovieRepository.cs
@@ -2,18 +2,25 @@
 using ErsatzTV.Core;
 using ErsatzTV.Core.Domain;
 using ErsatzTV.Core.Emby;
+using ErsatzTV.Core.Errors;
 using ErsatzTV.Core.Interfaces.Repositories;
 using ErsatzTV.Core.Metadata;
 using ErsatzTV.Infrastructure.Extensions;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
 
 namespace ErsatzTV.Infrastructure.Data.Repositories;
 
 public class EmbyMovieRepository : IEmbyMovieRepository
 {
     private readonly IDbContextFactory<TvContext> _dbContextFactory;
+    private readonly ILogger<EmbyMovieRepository> _logger;
 
-    public EmbyMovieRepository(IDbContextFactory<TvContext> dbContextFactory) => _dbContextFactory = dbContextFactory;
+    public EmbyMovieRepository(IDbContextFactory<TvContext> dbContextFactory, ILogger<EmbyMovieRepository> logger)
+    {
+        _dbContextFactory = dbContextFactory;
+        _logger = logger;
+    }
 
     public async Task<List<EmbyItemEtag>> GetExistingMovies(EmbyLibrary library)
     {
@@ -178,6 +185,11 @@ public class EmbyMovieRepository : IEmbyMovieRepository
     {
         try
         {
+            if (await MediaItemRepository.MediaFileAlreadyExists(movie, dbContext, _logger))
+            {
+                return new MediaFileAlreadyExists();
+            }
+
             // blank out etag for initial save in case other updates fail
             string etag = movie.Etag;
             movie.Etag = string.Empty;